// types/menu.ts
enum Course {
	SOUP = 'soup',
	MAIN = 'main',
	DESSERT = 'dessert',
}

enum DishFeature {
	APPLE = 'apple',
	ARUGULA = 'arugula',
	ASPARAGUS = 'asparagus',
	AVOCADO = 'avocado',
	BACON = 'bacon',
	BAGUETTE = 'baguette',
	BAKED = 'baked',
	BASIL = 'basil',
	BEAN = 'bean',
	BEEF = 'beef',
	BEETROOT = 'beetroot',
	BLACK_PEPPER = 'black_pepper',
	BOAR = 'boar',
	BOILED = 'boiled',
	BREAD = 'bread',
	BROCCOLI = 'broccoli',
	BUTTER = 'butter',
	CABBAGE = 'cabbage',
	CAPER = 'caper',
	CARAWAY = 'caraway',
	CARROT = 'carrot',
	CAULIFLOWER = 'cauliflower',
	CELERY = 'celery',
	CHEESE = 'cheese',
	CHICKEN = 'chicken',
	CHICKPEA = 'chickpea',
	CHILLI = 'chilli',
	CHIVE = 'chive',
	CITRUS = 'citrus',
	COCONUT = 'coconut',
	CORN = 'corn',
	CRANBERRY = 'cranberry',
	CREAM = 'cream',
	CROUTON = 'crouton',
	CUCUMBER = 'cucumber',
	CURD = 'curd',
	DESSERT = 'dessert',
	DILL = 'dill',
	DRESSING = 'dressing',
	DRIED = 'dried',
	DUCK = 'duck',
	DUMPLING = 'dumpling',
	EGG = 'egg',
	EGGPLANT = 'eggplant',
	FILLED = 'filled',
	FILLET = 'fillet',
	FISH = 'fish',
	FRESH = 'fresh',
	FRIED = 'fried',
	FRIES = 'fries',
	GARLIC = 'garlic',
	GINGER = 'ginger',
	GNOCCHI = 'gnocchi',
	GOULASH = 'goulash',
	GRILLED = 'grilled',
	HALUSKY = 'halusky',
	HAM = 'ham',
	HAMBURGER = 'hamburger',
	HERB = 'herb',
	HORSERADISH = 'horseradish',
	JUICE = 'juice',
	KARBANATEK = 'karbanatek',
	LAMB = 'lamb',
	LARD = 'lard',
	LEEK = 'leek',
	LEMON = 'lemon',
	LENTIL = 'lentil',
	LIME = 'lime',
	LIVER = 'liver',
	MARJORAM = 'marjoram',
	MASHED = 'mashed',
	MAYONNAISE = 'mayonnaise',
	MEAT = 'meat',
	MILK = 'milk',
	MINCED = 'minced',
	MUSHROOM = 'mushroom',
	NOODLES = 'noodles',
	NUT = 'nut',
	OIL = 'oil',
	ONION = 'onion',
	PARSLEY = 'parsley',
	PASTA = 'pasta',
	PEA = 'pea',
	PEANUT = 'peanut',
	PEPPER = 'pepper',
	PESTO = 'pesto',
	PIZZA = 'pizza',
	PORK = 'pork',
	POTATO = 'potato',
	PUREE = 'puree',
	RABBIT = 'rabbit',
	RICE = 'rice',
	SALAD = 'salad',
	SALMON = 'salmon',
	SAUCE = 'sauce',
	SAUSAGE = 'sausage',
	SCHNITZEL = 'schnitzel',
	SESAME = 'sesame',
	SKEWER = 'skewer',
	SMOKED = 'smoked',
	SOUP = 'soup',
	SPICE = 'spice',
	SPICY = 'spicy',
	SPINACH = 'spinach',
	STEAK = 'steak',
	STRAWBERRY = 'strawberry',
	SVICKOVA = 'svickova',
	TOAST = 'toast',
	TOMATO = 'tomato',
	TORTILLA = 'tortilla',
	TURKEY = 'turkey',
	VEGETABLE = 'vegetable',
	WHEAT = 'wheat',
	WHIPPED = 'whipped',
	WINE = 'wine',
	YOGURT = 'yogurt',
	ZUCCHINI = 'zucchini',
}

interface DailyMenuOffer {
	type: string
	price: number | number[]
}

interface Dish extends DailyMenuOffer {
	type: 'dish'
	name: string
	course?: Course
	features?: DishFeature[]
}

interface Menu extends DailyMenuOffer {
	type: 'menu',
	dishes: Record<Course, Dish[]>
}

interface DailyMenu {
	offers: DailyMenuOffer[]
}

export {
	Course,
	DishFeature,
	Dish,
	Menu,
	DailyMenuOffer,
	DailyMenu,
}